dbb2_requirements.txt
# NBA Fantasy Basketball Platform - Python Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0

# NBA Data
nba-api==1.4.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2

# Date/Time
python-dateutil==2.8.2

# Environment Variables
python-dotenv==1.0.0

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# JSON/Data Validation
pydantic==2.5.0

# Utilities
pytz==2023.3